Dashboard CRUDs DONE

+++ lib/_dao/objectImage-dao.ts
@@ -1,21 +1,21 @@
 'use server';
 
 import postgres from 'postgres';
-import { uploadFile } from '../upload-actions';
+import { deleteFile, uploadFile } from '../upload-actions';
+import { ObjectImage } from '../_objects/objectImage';
 
 const sql = postgres(process.env.POSTGRES_URL!, { ssl: 'require'});
 
 export async function createObjectImage(file: File, parentType: string, parentId: string) {
     
-    const fileName = `${parentType}-${parentId}-${file.name}`;
+    const filename = `${parentType}-${parentId}-${file.name}`;
 
-    console.log(fileName);
     try{
       await sql`
-        INSERT INTO objectImages (objectId, objectType, fileName)
-        VALUES (${parentId}, ${parentType}, ${fileName})
+        INSERT INTO objectImages (objectId, objectType, filename)
+        VALUES (${parentId}, ${parentType}, ${filename})
       `;
-      await uploadFile(file, fileName);
+      await uploadFile(file, filename);
     }catch (error) {
       return {
         message: 'Database Error: Failed to Create ObjectImage.',
@@ -44,6 +44,49 @@ export async function fetchObjectImageById(id: string) {
   }
 }
 
+export async function fetchObjectImageByObjectTypeAndId(objectType: string, objectId: string) {
+  try {
+    const data = await sql<ObjectImage[]>`
+      SELECT
+        objectImages.id,
+        objectImages.filename,
+        objectImages.objectId,
+        objectImages.objectType
+      FROM objectImages
+      WHERE objectImages.objectid = ${objectId} AND objectImages.objectType = ${objectType};
+    `;
+
+    const objectImage = data.map((objectImage) => ({...objectImage}));
+
+    return objectImage;
+  } catch (error) {
+    console.error('Database Error:', error);
+    throw new Error('Failed to fetch objectImages.');
+  }
+}
+
 export async function deleteObjectImage(id: string){
-    await sql`DELETE FROM objectImages WHERE id = ${id}`;
+  try{
+    const objectImage = await fetchObjectImageById(id);
+    if(objectImage){
+      await sql`DELETE FROM objectImages WHERE id = ${id}`;
+      deleteFile(objectImage.filename);
+    }
+  }catch(error){
+    console.error('Database Error:', error);
+    throw new Error('Failed to delete objectImage.');
+  }
+}
+
+export async function deleteObjectImageByObjectIdAndObjectType(objectId: string, objectType: string){
+  try{
+    const objectImages = await fetchObjectImageByObjectTypeAndId(objectType, objectId);
+    if(objectImages){
+      await sql`DELETE FROM objectImages WHERE objectid = ${objectId} AND objecttype = ${objectType}`;
+      objectImages.forEach((objectImage) => deleteFile(objectImage.filename))
+    }
+  }catch(error){
+    console.error('Database Error:', error);
+    throw new Error('Failed to delete objectImages.');
+  }
 }
